# apispec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/training/yolo/detection": {
      "post": {
        "tags": [
          "Training",
          "Training"
        ],
        "summary": "Yolo Training",
        "description": "Start YOLO training with the provided parameters.",
        "operationId": "yolo_training_training_yolo_detection_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YoloDetTrainingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/training/yolo/result": {
      "post": {
        "tags": [
          "Training",
          "Training"
        ],
        "summary": "Yolo Training Result",
        "description": "Handle the result of YOLO training.",
        "operationId": "yolo_training_result_training_yolo_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YoloHandlingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/validation/yolo/result": {
      "post": {
        "tags": [
          "Validation",
          "Validation"
        ],
        "summary": "Receive Result",
        "description": "Receive validation completion callback and persist to Mongo/MinIO.",
        "operationId": "receive_result_validation_yolo_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Result"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/IDE/codebase": {
      "get": {
        "tags": [
          "Code Editing",
          "Code Editor"
        ],
        "summary": "Get Codebase",
        "description": "Get codebase to edit.",
        "operationId": "get_codebase_IDE_codebase_get",
        "parameters": [
          {
            "name": "cid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/IDE/codebase/create": {
      "post": {
        "tags": [
          "Code Editing",
          "Code Editor"
        ],
        "summary": "Create Cb",
        "description": "Create a snapshot of the current training state.",
        "operationId": "create_cb_IDE_codebase_create_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_cb_IDE_codebase_create_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/IDE/codebase/update": {
      "post": {
        "tags": [
          "Code Editing",
          "Code Editor"
        ],
        "summary": "Update Cb",
        "description": "Update a snapshot of the current training state.",
        "operationId": "update_cb_IDE_codebase_update_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_cb_IDE_codebase_update_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_cb_IDE_codebase_create_post": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "request": {
            "$ref": "#/components/schemas/CodebaseCreateRequest"
          }
        },
        "type": "object",
        "required": [
          "data",
          "request"
        ],
        "title": "Body_create_cb_IDE_codebase_create_post"
      },
      "Body_update_cb_IDE_codebase_update_post": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "request": {
            "$ref": "#/components/schemas/CodebaseUpdateRequest"
          }
        },
        "type": "object",
        "required": [
          "data",
          "request"
        ],
        "title": "Body_update_cb_IDE_codebase_update_post"
      },
...
      "CodebaseCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "algorithm": {
            "type": "string",
            "title": "Algorithm"
          },
          "stage": {
            "type": "string",
            "title": "Stage",
            "description": "Stage of the training snapshot (e.g., 'Training', 'Validating', 'Optimizing')"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "algorithm",
          "stage",
          "task_type"
        ],
        "title": "CodebaseCreateRequest"
      },
      "CodebaseUpdateRequest": {
        "properties": {
          "cid": {
            "type": "string",
            "title": "Cid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "algorithm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Algorithm"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage",
            "description": "Stage of the training snapshot (e.g., 'Training', 'Validating', 'Optimizing')"
          },
          "task_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "cid",
          "name",
          "algorithm",
          "stage",
          "task_type"
        ],
        "title": "CodebaseUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "YoloDetTrainingParams": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "split_ratio": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Split Ratio"
          },
          "epochs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Epochs"
          },
          "batch": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch"
          },
          "imgsz": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imgsz"
          },
          "device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device"
          },
          "save_period": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Save Period"
          },
          "workers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workers"
          },
          "pretrained": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pretrained"
          },
          "optimizer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimizer"
          },
          "lr0": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lr0"
          },
          "lrf": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lrf"
          },
          "momentum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Momentum"
          },
          "weight_decay": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Decay"
          },
          "patience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patience"
          },
          "augment": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Augment"
          },
          "warmup_epochs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warmup Epochs"
          },
          "warmup_momentum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warmup Momentum"
          },
          "warmup_bias_lr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warmup Bias Lr"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "cache": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache"
          },
          "dropout": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dropout"
          },
          "label_smoothing": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label Smoothing"
          },
          "rect": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rect"
          },
          "resume": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume"
          },
          "amp": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amp"
          },
          "single_cls": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Single Cls"
          },
          "cos_lr": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cos Lr"
          },
          "close_mosaic": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Mosaic"
          },
          "overlap_mask": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlap Mask"
          },
          "mask_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mask Ratio"
          }
        },
        "type": "object",
        "title": "YoloDetTrainingParams"
      },
      "YoloDetTrainingRequest": {
        "properties": {
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "cid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cid"
          },
          "did": {
            "type": "string",
            "title": "Did"
          },
          "origin_tid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Tid"
          },
          "parameters": {
            "$ref": "#/components/schemas/YoloDetTrainingParams"
          }
        },
        "type": "object",
        "required": [
          "pid",
          "did",
          "parameters"
        ],
        "title": "YoloDetTrainingRequest"
      },
...
      "YoloHandlingRequest": {
        "properties": {
          "workdir": {
            "type": "string",
            "title": "Workdir"
          },
          "result": {
            "$ref": "#/components/schemas/YoloTrainingResult"
          }
        },
        "type": "object",
        "required": [
          "workdir",
          "result"
        ],
        "title": "YoloHandlingRequest"
      },
      "YoloTrainingResult": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "origin_tid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Tid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "workdir": {
            "type": "string",
            "title": "Workdir"
          },
          "started_time": {
            "type": "string",
            "title": "Started Time"
          },
          "completed_time": {
            "type": "string",
            "title": "Completed Time"
          },
          "parameters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameters"
          },
          "did": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Did"
          },
          "cid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cid"
          },
          "artifacts_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts Path"
          },
          "error_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Details"
          }
        },
        "type": "object",
        "required": [
          "uid",
          "pid",
          "status",
          "task_type",
          "workdir",
          "started_time",
          "completed_time"
        ],
        "title": "YoloTrainingResult"
      }
    }
  }
}